refactored file names to fit more data/added tangible magic buttons
