+++ src/TangibleTypePage.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+function TangibleTypePage () {}
+
+export default TangibleTypePage
